+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/process/M_ShipperTransportation_AddShipments.java
@@ -2,7 +2,7 @@ package de.metas.handlingunits.shipping.process;
 
 
 import com.google.common.collect.ImmutableList;
-import de.metas.handlingunits.transportation.InOutToTransportationOrderService;
+import de.metas.handlingunits.shipping.InOutToTransportationOrderService;
 import de.metas.inout.IInOutDAO;
 import de.metas.inout.InOutId;
 import de.metas.process.IProcessPrecondition;
